requirements.txt
Django==2.1.12
djangorestframework==3.10.3
django-cors-middleware==1.3.1
psycopg2==2.8.3
PyYAML==5.1.2
django-rest-swagger==2.2.0
pytest==5.1.3
pytest-django==3.5.1
pytest-cov==2.7.1
drf-jwt==1.13.3
pactman
locustio
geventhttpclient
drf-serializer-cache
gunicorn
requests